fix: 将 stats_models 表 ID 类型从 int 改为 int64 修复 32 位平台溢出崩溃 (#41)
buildModelID 使用 FNV-1a 64 位哈希生成 ID，最大值可达 2^63-1，
在 32 位平台（armv7/386）上 int 仅 32 位导致溢出。
将 StatsModel.ID、cache key、dirty map 及相关函数签名统一改为 int64。

+++ internal/op/stats/cache.go
@@ -71,7 +71,7 @@ func RefreshCache(ctx context.Context) error {
 
 	modelCache.Clear()
 	modelCacheNeedUpdateLock.Lock()
-	modelCacheNeedUpdate = make(map[int]struct{})
+	modelCacheNeedUpdate = make(map[int64]struct{})
 	modelCacheNeedUpdateLock.Unlock()
 	for _, v := range loadedModels {
 		modelCache.Set(v.ID, v)
@@ -102,4 +102,3 @@ func RefreshCache(ctx context.Context) error {
 
 	return nil
 }
-
